requirements.txt
pyopenjtalk-plus
tqdm
nltk
jaconv